requirements.txt
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
tensorboard==2.16.2
matplotlib==3.9.2
streamlit==1.39.0